refactor: extracted successComponent
