+++ packages/tradershub/src/routes/TradersHubRoute/TradersHubRoute.tsx
@@ -9,6 +9,7 @@ import {
 } from '../../components';
 import { CTraderList } from '../../features/cfd/components/CTraderList';
 import { OtherCFDPlatformsList } from '../../features/cfd/components/OtherCFDPlatformsList';
+import { MT5PasswordModal } from '../../features/cfd/modals/MT5PasswordModal';
 
 const TradersHubRoute = () => {
     const { isMobile } = useBreakpoint();
@@ -42,6 +43,7 @@ const TradersHubRoute = () => {
                 <TotalAssets />
             </div>
             <OptionsAndMultipliersSection />
+            <MT5PasswordModal marketType={'synthetic'} platform={'mt5'} />
 
             <div className='border-solid p-1200 rounded-1200 border-xs border-opacity-black-100'>
                 <div className='pb-1200'>
